obj relational mapping

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,28 +1,24 @@
 #!/usr/bin/python3
-"""script that deletes all State objects with a
-name containing the letter a from the database
-hbtn_0e_6_usa
 """
-
-import sys
+This script is  deleting  all State objects
+with a name containing the letter `a`
+from the database `hbtn_0e_6_usa`.
+"""
+from sys import argv
 from model_state import Base, State
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy import create_engine
-
+from sqlalchemy import (create_engine)
+from sqlalchemy.orm import Session
 
 if __name__ == "__main__":
-    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
-                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.
+                           format(argv[1], argv[2], argv[3]),
                            pool_pre_ping=True)
-    Session = sessionmaker(bind=engine)
-    session = Session()
     Base.metadata.create_all(engine)
 
-    states = session.query(State).filter(State.name.contains('a')).all()
-
-    for state in states:
-        session.delete(state)
+    session = Session(engine)
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
+            session.delete(state)
 
     session.commit()
-
-    session.close()
